+++ django/applications/catmaid/static/js/widgets/analyze_arbor.js
@@ -164,10 +164,7 @@ AnalyzeArbor.prototype.init = function() {
       "bServerSide": false, // Enable sorting locally, and prevent sorting from calling the fnServerData to reload the table -- an expensive and undesirable operation.
       "bAutoWidth": false,
       "iDisplayLength": -1,
-      "aLengthMenu": [
-        [-1, 10, 100, 200],
-        ["All", 10, 100, 200]
-      ],
+      "aLengthMenu": [CATMAID.pageLengthOptions, CATMAID.pageLengthLabels],
       //"aLengthChange": false,
       "bJQueryUI": true,
       "aoColumns": [{bSearchable: true, bSortable: true}].concat((function() {